fix(web): avoid react-native pager import leakage from EarnHome
